api/crewai_agents/requirements.txt
crewai==0.5.0
python-dotenv==1.1.1
requests>=2.32.0
regex>=2024.4.16
